+++ packages/expo-sensors/src/ExpoLightSensor.web.ts
@@ -1,0 +1,10 @@
+export default {
+  get name(): string {
+    return 'ExpoLightSensor';
+  },
+  async isAvailableAsync(): Promise<boolean> {
+    return false;
+  },
+  startObserving() {},
+  stopObserving() {},
+};