+++ packages/expo-sensors/src/index.ts
@@ -7,5 +7,6 @@ export { default as DeviceSensor } from './DeviceSensor';
 export { default as Gyroscope } from './Gyroscope';
 export { default as Magnetometer } from './Magnetometer';
 export { default as MagnetometerUncalibrated } from './MagnetometerUncalibrated';
+export { default as LightSensor, LightSensorMeasurement } from './LightSensor';
 export { ThreeAxisMeasurement } from './ThreeAxisSensor';
 export { Pedometer };